[UPDATE] 댓글 및 게시글 삭제 로직 수정
- 게시자 id와 현재 로그인 된 유저 id가 동일할 경우 삭제하도록 수정
- 삭제 페이지 여러번 방문하는 오류 수정

+++ pages/board/update-comment/delete.js
@@ -1,6 +1,9 @@
 import { useEffect } from 'react';
 import Router, { useRouter } from 'next/router';
-import { useDispatch } from '../../../components/Contexts/context';
+import {
+  useDispatch,
+  useReducerState,
+} from '../../../components/Contexts/context';
 import DeleteTemplate from '../../../components/Templates/DeleteTemplate';
 import Head from 'next/head';
 import { deleteComment } from '../../../utils/updateFunc/comment/deleteComment';
@@ -10,20 +13,23 @@ export default function Delete() {
     boardtype: boardType,
     postid: postId,
     commentid: commentId,
+    authorid: authorId,
   } = useRouter().query;
   const dispatch = useDispatch();
+  const { userId } = useReducerState().user;
 
   useEffect(() => {
     dispatch({
       type: 'initiate',
       nowPage: 'delete',
     });
-    if (boardType || postId || commentId) {
+    if (boardType && postId && commentId && authorId === userId) {
       deleteComment({ boardType, postId, commentId, dispatch });
     } else {
       Router.push('/404');
     }
-  }, [boardType, commentId, dispatch, postId]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <>
